app.py: Need to pass in port number as 'int'
When we configured an explicite db_port in mediagoblin.ini, paster would
crash claiming that the port number must be an int. Given that we don't
have a "get_conf_int()" function or something similar (yet?), simply
convert the port number to int before passing it to the mongo Connection
instance.

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>